image_augmentor/requirement.txt
numpy
opencv-python
Pillow
PyYAML
torchvision
torch          